+++ test/Numbers/parseNumber.ipp
@@ -131,7 +131,7 @@ using ArduinoJson::Internals::parseJsonNumber;
 #undef CATCH_INTERNAL_LINEINFO
 #define CATCH_INTERNAL_LINEINFO source_line_info
 
-#define __CATCH_LOCAL_LINEINFO ::Catch::SourceLineInfo source_line_info = ::Catch::SourceLineInfo(__builtin_FILE(), __builtin_LINE())
+#define __CATCH_LOCAL_LINEINFO ::Catch::SourceLineInfo source_line_info = ::Catch::SourceLineInfo(__builtin_FILE(), static_cast<size_t>(__builtin_LINE()))
 #define __CATCH_NESTED_LINEINFO ::Catch::SourceLineInfo source_line_info
 
 inline void expectNumber(const char* s, __CATCH_LOCAL_LINEINFO) {
